+++ src/PiProxyGuard.Infrastructure/Migrations/20260613165312_InitUpdate.cs
@@ -1,0 +1,22 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace PiProxyGuard.Infrastructure.Migrations
+{
+    /// <inheritdoc />
+    public partial class InitUpdate : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+
+        }
+    }
+}